+++ eugene/src/SensorPlugins/Riken/Sensor.Riken.h
@@ -39,13 +39,13 @@ class SensorRiken : public Sensor
   int RAFL_A_Traiter;           // Index du RIKEN en cours
   int RAFL_A_Traiter_Remenber;
   int StrandRespect;
-  int MIN_EST_DIFF; // default = 100;
-  int MAX_OVERLAP; // default = 60;
-  int MAX_RIKEN_LENGTH; // default = 60000;
+  int MIN_EST_DIFF;         // default = 100;
+  int MAX_OVERLAP;          // default = 60;
+  int MAX_RIKEN_LENGTH;     // default = 60000;
   int MAX_RIKEN_EST_LENGTH; // default  = 3000;
-  int MIN_RIKEN_LENGTH; // default  = 120; // 2* riken overlap (60)
+  int MIN_RIKEN_LENGTH;     // default  = 120; // 2* riken overlap (60)
   int MIN_RIKEN_EST_LENGTH; // default  = 10;
-  REAL RAFLPenalty; // default  = -120;
+  double RAFLPenalty;       // default  = -120;
 
  public:
   SensorRiken   (int n, DNASeq *X);
